test(daemon): parameterize multiplayer suite over network; run on OCapN
Extract the 9 multi-daemon invite/accept/value-exchange/restart/
partition/three-party/sub-invitation-chain/agent-ring tests from
`invite-retention.test.js` into a shared `runMultiplayerSuite({ test,
network })` in `_multiplayer-suite.js`, parameterized over a
`NetworkSpec` (listen-address pet name, module path, `@nets/<key>`
destination). The TCP entrypoint becomes a 25-line wrapper that
invokes the suite with `tcpNetwork`; a new `invite-retention-ocapn.
test.js` invokes the same suite with `ocapnNetwork`.

This is the daemon-level coverage the design doc's Test Plan calls
for: two daemons forked end-to-end, the full MULTIPLAYER.md flow
(invite, accept, send a value, adopt it, request, resolve) actually
running over OCapN-Noise sessions instead of CapTP. Before this,
OCapN coverage was three in-process tests in `networks-ocapn.test.js`
that mocked `@agent` powers and never went through `makePeer`,
`provideRemoteControl`, or the daemon's invite/accept formulas.

All 9 multiplayer tests pass on both transports. The OCapN-Noise
transport conforms cleanly through the daemon's `getAllNetworkAddresses`
/ `makePeer` / `EndoGreeter` / `EndoGateway` / `RemoteControl` stack;
no daemon-side changes were needed to make the suite green over OCapN.
