+++ packages/daemon/test/invite-retention.test.js
@@ -1,23 +1,15 @@
 // @ts-check
-/* global process, setTimeout */
+/* global process */
 
 // Establish a perimeter:
 // eslint-disable-next-line import/order
 import '@endo/init/debug.js';
 
 import baseTest from 'ava';
-import os from 'os';
-import url from 'url';
-import path from 'path';
-import { E } from '@endo/far';
-import { makePromiseKit } from '@endo/promise-kit';
-import { start, stop, restart, purge, makeEndoClient } from '../index.js';
-import { parseId } from '../src/formula-identifier.js';
-import { idFromLocator } from '../src/locator.js';
-import { makeDaemonDatabase } from '../src/daemon-database-node.js';
+import { runMultiplayerSuite, tcpNetwork } from './_multiplayer-suite.js';
 
-// Multi-daemon retention tests rely on the TCP network module that
-// is loaded via makeUnconfined (a Node-only path).  Skip the whole
+// Multi-daemon retention tests rely on a network module that is
+// loaded via makeUnconfined (a Node-only path).  Skip the whole
 // suite on the bare Rust supervisor (test:rust without
 // ENDO_NODE_WORKER_BIN).
 const skipNoNodeWorker =
@@ -30,573 +22,4 @@ const test = skipNoNodeWorker
     })
   : baseTest;
 
-const dirname = url.fileURLToPath(new URL('..', import.meta.url)).toString();
-
-// ── Helpers ──────────────────────────────────────────────────────
-
-const MAX_CONFIG_DIR_LENGTH = 80;
-let configPathId = 0;
-
-const getConfigDirectoryName = (testTitle, configNumber) => {
-  const cleanTitle = testTitle
-    .replace(/[^a-zA-Z0-9-]/g, '-')
-    .toLowerCase()
-    .slice(0, 50);
-  const defaultPath = `${cleanTitle}`;
-  const basePath =
-    defaultPath.length <= MAX_CONFIG_DIR_LENGTH
-      ? defaultPath
-      : defaultPath.slice(0, MAX_CONFIG_DIR_LENGTH);
-  const testId = String(configPathId).padStart(4, '0');
-  const configId = String(configNumber).padStart(2, '0');
-  const configSubDirectory = `${basePath}#${testId}-${configId}`;
-  configPathId += 1;
-  return configSubDirectory;
-};
-
-/** @param  {...string} root */
-const makeConfig = (...root) => {
-  // Use a short path for the socket to stay within the ~108 char
-  // Unix socket path limit.  CI checkout paths can be long.
-  // The last root segment includes a unique test/config ID suffix.
-  const tag = root.join('-').slice(-40);
-  const shortSock = path.join(os.tmpdir(), `endo-${tag}.sock`);
-  return {
-    statePath: path.join(dirname, ...root, 'state'),
-    ephemeralStatePath: path.join(dirname, ...root, 'run'),
-    cachePath: path.join(dirname, ...root, 'cache'),
-    sockPath:
-      process.platform === 'win32'
-        ? String.raw`\\?\pipe\endo-${root.join('-')}-test.sock`
-        : shortSock,
-    address: '127.0.0.1:0',
-    pets: new Map(),
-    values: new Map(),
-  };
-};
-
-const prepareConfig = async (t, { gcEnabled = true } = {}) => {
-  const { reject: cancel, promise: cancelled } = makePromiseKit();
-  cancelled.catch(() => {});
-  const config = {
-    ...makeConfig('tmp', getConfigDirectoryName(t.title, t.context.length)),
-    gcEnabled,
-  };
-  await purge(config);
-  await start(config);
-  const contextObj = { cancel, cancelled, config };
-  t.context.push(contextObj);
-  return { ...contextObj };
-};
-
-const makeHost = async (config, cancelled) => {
-  const { getBootstrap, closed } = await makeEndoClient(
-    'client',
-    config.sockPath,
-    cancelled,
-  );
-  closed.catch(() => {});
-  const bootstrap = getBootstrap();
-  return { host: E(bootstrap).host() };
-};
-
-const prepareHostWithGcAndNetwork = async t => {
-  const { cancel, cancelled, config } = await prepareConfig(t, {
-    gcEnabled: true,
-  });
-  const { host } = await makeHost(config, cancelled);
-
-  await E(host).storeValue('127.0.0.1:0', 'tcp-listen-addr');
-  const servicePath = path.join(dirname, 'src', 'networks', 'tcp-netstring.js');
-  const serviceLocation = url.pathToFileURL(servicePath).href;
-  const network = await E(host).makeUnconfined('@main', serviceLocation, {
-    powersName: '@agent',
-    resultName: 'test-network',
-  });
-  await network;
-  await E(host).move(['test-network'], ['@nets', 'tcp']);
-
-  return { host, config, cancel, cancelled };
-};
-
-const openTestDb = statePath => {
-  return makeDaemonDatabase({
-    statePath,
-    ephemeralStatePath: '',
-    cachePath: '',
-    sockPath: '',
-  });
-};
-
-const formulaExistsInDb = (statePath, id) => {
-  const { number } = parseId(id);
-  return openTestDb(statePath).hasFormula(number);
-};
-
-/**
- * Wait for a condition to become true, polling at intervals.
- *
- * @param {() => boolean | Promise<boolean>} check
- * @param {{ timeoutMs?: number, intervalMs?: number }} [opts]
- */
-const waitForCondition = async (check, opts = {}) => {
-  const { timeoutMs = 5000, intervalMs = 100 } = opts;
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() < deadline) {
-    // eslint-disable-next-line no-await-in-loop
-    if (await check()) return;
-    // eslint-disable-next-line no-await-in-loop
-    await new Promise(resolve => setTimeout(resolve, intervalMs));
-  }
-  throw new Error(`Condition not met within ${timeoutMs}ms`);
-};
-
-// ── Lifecycle ────────────────────────────────────────────────────
-
-test.beforeEach(t => {
-  t.context = [];
-});
-
-test.afterEach.always(async t => {
-  const configs = /** @type {any[]} */ (t.context);
-  await Promise.allSettled(configs.map(({ config }) => stop(config)));
-  for (const { cancel, cancelled } of configs) {
-    cancelled.catch(() => {});
-    cancel(Error('teardown'));
-  }
-});
-
-// ── Tests ────────────────────────────────────────────────────────
-
-test.serial(
-  'deleting invited guest pet name collects guest formulas',
-  async t => {
-    const { host, config } = await prepareHostWithGcAndNetwork(t);
-
-    // Create a guest.
-    await E(host).provideGuest('my-guest', { agentName: 'my-agent' });
-
-    // Verify the guest formula exists.
-    const guestId = await E(host).identify('my-agent');
-    const handleId = await E(host).identify('my-guest');
-    t.true(formulaExistsInDb(config.statePath, guestId), 'guest exists');
-    t.true(formulaExistsInDb(config.statePath, handleId), 'handle exists');
-
-    // Remove both the pet name and the pin.
-    await E(host).remove('my-agent');
-    await E(host).remove('my-guest');
-
-    // The guest and handle should be collected.
-    await waitForCondition(
-      () =>
-        !formulaExistsInDb(config.statePath, guestId) &&
-        !formulaExistsInDb(config.statePath, handleId),
-    );
-    t.false(formulaExistsInDb(config.statePath, guestId), 'guest collected');
-    t.false(formulaExistsInDb(config.statePath, handleId), 'handle collected');
-  },
-);
-
-test.serial('invited guest retains values shared through mail', async t => {
-  const { host: hostA, config: configA } = await prepareHostWithGcAndNetwork(t);
-  const { host: hostB } = await prepareHostWithGcAndNetwork(t);
-
-  // Establish invite/accept.
-  const invitation = await E(hostA).invite('bob');
-  const invitationLocator = await E(invitation).locate();
-  await E(hostB).accept(invitationLocator, 'alice');
-
-  // Create a value on A.
-  await E(hostA).evaluate('@main', '"shared-value"', [], [], ['shared']);
-  const sharedLocator = await E(hostA).locate('shared');
-  const sharedId = idFromLocator(sharedLocator);
-
-  // Send it to bob.
-  await E(hostA).send('bob', ['Here'], ['shared'], ['shared']);
-
-  // The value should remain alive on A (bob references it via mail).
-  t.true(
-    formulaExistsInDb(configA.statePath, sharedId),
-    'shared value exists on A',
-  );
-
-  // B receives the message.
-  const messages = /** @type {unknown[]} */ (await E(hostB).listMessages());
-  t.true(messages.length > 0, 'B received a message');
-});
-
-test.serial(
-  'deleting invited guest and pin collects guest formulas',
-  async t => {
-    const { host: hostA, config: configA } =
-      await prepareHostWithGcAndNetwork(t);
-    const { host: hostB } = await prepareHostWithGcAndNetwork(t);
-
-    // Establish invite/accept.
-    const invitation = await E(hostA).invite('bob');
-    const invitationLocator = await E(invitation).locate();
-    await E(hostB).accept(invitationLocator, 'alice');
-
-    // The guest handle is stored under 'bob' and also pinned in
-    // '@pins/guest-bob'. Both must be removed for collection.
-    const bobNames = await E(hostA).list();
-    t.true(bobNames.includes('bob'), 'bob exists on A after accept');
-
-    // Get the local formula ID stored in A's pet store for 'bob'.
-    const bobId = await E(hostA).identify('bob');
-    t.truthy(bobId, 'bob has an identity on A');
-
-    // The pinned guest formula should also exist.
-    const pinnedId = await E(hostA).identify('@pins', 'guest-bob');
-    t.truthy(pinnedId, 'guest-bob pin exists');
-
-    // Remove both references.
-    await E(hostA).remove('bob');
-    await E(hostA).remove('@pins', 'guest-bob');
-
-    // Wait for GC to collect the handle formula.
-    await waitForCondition(() => !formulaExistsInDb(configA.statePath, bobId), {
-      timeoutMs: 5000,
-    });
-    t.false(
-      formulaExistsInDb(configA.statePath, bobId),
-      'bob handle formula collected after removing all references',
-    );
-  },
-);
-
-test.serial('partition does not prevent local value release', async t => {
-  const { host: hostA, config: configA } = await prepareHostWithGcAndNetwork(t);
-  const { host: hostB, config: configB } = await prepareHostWithGcAndNetwork(t);
-
-  // Establish invite/accept.
-  const invitation = await E(hostA).invite('bob');
-  const invitationLocator = await E(invitation).locate();
-  await E(hostB).accept(invitationLocator, 'alice');
-
-  // Create a local-only value on A (not shared with B).
-  await E(hostA).storeValue({ local: true }, 'local-only');
-  const localLocator = await E(hostA).locate('local-only');
-  const localId = idFromLocator(localLocator);
-  t.true(formulaExistsInDb(configA.statePath, localId), 'local value exists');
-
-  // Simulate partition: stop B's daemon.
-  await stop(configB);
-
-  // Release the local value on A while partitioned.
-  await E(hostA).remove('local-only');
-
-  // GC should still collect it — partition doesn't prevent local GC.
-  await waitForCondition(() => !formulaExistsInDb(configA.statePath, localId));
-  t.false(
-    formulaExistsInDb(configA.statePath, localId),
-    'local value collected during partition',
-  );
-});
-
-test.serial(
-  'value shared with remote peer survives local release during partition',
-  async t => {
-    const { host: hostA, config: configA } =
-      await prepareHostWithGcAndNetwork(t);
-    const { host: hostB, config: configB } =
-      await prepareHostWithGcAndNetwork(t);
-
-    // Bidirectional peer exchange.
-    await E(hostA).addPeerInfo(await E(hostB).getPeerInfo());
-    await E(hostB).addPeerInfo(await E(hostA).getPeerInfo());
-
-    // Create a value on A and share it with B.
-    await E(hostA).evaluate('@main', '"important"', [], [], ['data']);
-    const dataLocator = await E(hostA).locate('data');
-    const dataId = idFromLocator(dataLocator);
-
-    // B stores and accesses the shared value.
-    await E(hostB).storeLocator(['from-a'], dataLocator);
-    const val = await E(hostB).lookup(['from-a']);
-    t.is(val, 'important');
-
-    // Simulate partition: stop B.
-    await stop(configB);
-
-    // A removes its local name for the value while B is partitioned.
-    await E(hostA).remove('data');
-
-    // The formula should still exist on A because B's retention set
-    // (from before partition) should keep it alive.
-    // NOTE: With the current architecture where remote values aren't
-    // stored locally, the retention set may be empty. The formula
-    // is only retained if it has other local references.
-    // For now, verify the formula is collected (no remote retention).
-    await waitForCondition(
-      () => !formulaExistsInDb(configA.statePath, dataId),
-      { timeoutMs: 3000 },
-    ).catch(() => {});
-
-    // Record whether it was collected (expected: yes, until per-agent
-    // formula IDs enable meaningful retention sets).
-    const collected = !formulaExistsInDb(configA.statePath, dataId);
-    t.pass(
-      collected
-        ? 'formula collected (retention not yet populated — expected for now)'
-        : 'formula retained by remote retention set',
-    );
-  },
-);
-
-test.serial('invite/accept works across restart', async t => {
-  const {
-    host: hostA,
-    config: configA,
-    cancelled: cancelledA,
-  } = await prepareHostWithGcAndNetwork(t);
-  const { host: hostB } = await prepareHostWithGcAndNetwork(t);
-
-  // Establish invite/accept.
-  const invitation = await E(hostA).invite('bob');
-  const invitationLocator = await E(invitation).locate();
-  await E(hostB).accept(invitationLocator, 'alice');
-
-  // Send a message.
-  await E(hostA).send('bob', ['Before restart'], [], []);
-
-  // Verify B received it.
-  const messagesBefore = await E(hostB).listMessages();
-  t.true(
-    messagesBefore.some(m => m.strings && m.strings[0] === 'Before restart'),
-    'B received message before restart',
-  );
-
-  // Restart A.
-  await restart(configA);
-  const { host: hostA2 } = await makeHost(configA, cancelledA);
-
-  // A should still know bob after restart.
-  const bobNames = await E(hostA2).list();
-  t.true(bobNames.includes('bob'), 'bob persists across restart');
-});
-
-test.serial('three-party invite with partition and recovery', async t => {
-  const { host: hostA } = await prepareHostWithGcAndNetwork(t);
-  const { host: hostB, config: configB } = await prepareHostWithGcAndNetwork(t);
-  const { host: hostC } = await prepareHostWithGcAndNetwork(t);
-
-  // A invites B and C.
-  const invB = await E(hostA).invite('bob');
-  const invC = await E(hostA).invite('carol');
-  await E(hostB).accept(await E(invB).locate(), 'alice');
-  await E(hostC).accept(await E(invC).locate(), 'alice');
-
-  // A sends to both.
-  await E(hostA).evaluate('@main', '"for-all"', [], [], ['shared']);
-  await E(hostA).send('bob', ['Hi Bob'], ['shared'], ['shared']);
-  await E(hostA).send('carol', ['Hi Carol'], ['shared'], ['shared']);
-
-  // Both receive.
-  const bobMsgs = await E(hostB).listMessages();
-  const carolMsgs = await E(hostC).listMessages();
-  t.true(
-    bobMsgs.some(m => m.strings?.[0] === 'Hi Bob'),
-    'B received message',
-  );
-  t.true(
-    carolMsgs.some(m => m.strings?.[0] === 'Hi Carol'),
-    'C received message',
-  );
-
-  // Partition B.
-  await stop(configB);
-
-  // A can still communicate with C while B is partitioned.
-  await E(hostA).evaluate('@main', '"after-partition"', [], [], ['new-val']);
-  await E(hostA).send('carol', ['Still here'], ['new-val'], ['new-val']);
-
-  const carolMsgs2 = await E(hostC).listMessages();
-  t.true(
-    carolMsgs2.some(m => m.strings?.[0] === 'Still here'),
-    'C received message during B partition',
-  );
-});
-
-// On both inviter and acceptor sides, after invite/accept:
-//   identify(guestName)            -> the *remote* handle id
-//                                     (not present in the local DB)
-//   identify('@pins', `guest-${guestName}`)
-//                                  -> the *local* guest handle id
-//                                     (the only formula whose
-//                                     collection is observable in
-//                                     this daemon's DB).
-// Both names must be removed for the local handle to become
-// unreachable; only the pin id is meaningful for asserting local
-// collection.
-
-// Sub-invitation chain: A->B->C. B is the inviter for C, so the
-// invitation/handle formulas for the C side originate on B's daemon.
-// When C drops their local references for the chain, C's local
-// guest-handle (held under '@pins/guest-bob-from-B') should collect
-// on C's daemon. A's roots for the original A->B invitation must
-// remain reachable: nothing C does on its own daemon should
-// invalidate A's local pin.
-test.serial(
-  'sub-invitation chain (A->B->C) collects C-side resources after C release',
-  async t => {
-    t.timeout(60000);
-    const { host: hostA, config: configA } =
-      await prepareHostWithGcAndNetwork(t);
-    const { host: hostB, config: configB } =
-      await prepareHostWithGcAndNetwork(t);
-    const { host: hostC, config: configC } =
-      await prepareHostWithGcAndNetwork(t);
-
-    // A invites B. B accepts as 'alice-from-A'. A's pet 'bob' now
-    // points to A's local handle for B.
-    const invAtoB = await E(hostA).invite('bob');
-    await E(hostB).accept(await E(invAtoB).locate(), 'alice-from-A');
-
-    // B invites C, C accepts. This is the sub-invitation: B is the
-    // inviter (creating an invitation formula on B's daemon), and C
-    // accepts to obtain a local guest handle on C's daemon.
-    const invBtoC = await E(hostB).invite('carol');
-    await E(hostC).accept(await E(invBtoC).locate(), 'bob-from-B');
-
-    // The pin id is the formula on C's daemon for C's local guest
-    // handle. The pet name 'bob-from-B' on C points to a remote
-    // handle id (B's handle) and is not in C's DB.
-    const cPinId = await E(hostC).identify('@pins', 'guest-bob-from-B');
-    t.truthy(cPinId, 'C pinned the guest handle');
-    t.true(
-      formulaExistsInDb(configC.statePath, cPinId),
-      'C local guest handle exists pre-release',
-    );
-
-    // Capture A's pin (the local handle for the original A->B
-    // invitation) so we can assert it stays reachable.
-    const aPinId = await E(hostA).identify('@pins', 'guest-bob');
-    t.truthy(aPinId, "A has '@pins/guest-bob'");
-    t.true(
-      formulaExistsInDb(configA.statePath, aPinId),
-      "A's 'guest-bob' pin exists",
-    );
-
-    // C drops both references for the chain it joined. Only after
-    // *both* the pet name and the pin are gone can the local handle
-    // collect (the pin keeps the local handle alive under @pins).
-    await E(hostC).remove('bob-from-B');
-    await E(hostC).remove('@pins', 'guest-bob-from-B');
-
-    // C's local guest handle should collect on C's daemon.
-    await waitForCondition(
-      () => !formulaExistsInDb(configC.statePath, cPinId),
-      { timeoutMs: 8000 },
-    );
-    t.false(
-      formulaExistsInDb(configC.statePath, cPinId),
-      'C local guest handle collected after C dropped both references',
-    );
-
-    // A's pin for 'bob' must still be reachable: nothing C does on
-    // its own daemon should affect A's local roots in the chain.
-    t.true(
-      formulaExistsInDb(configA.statePath, aPinId),
-      "A's 'guest-bob' pin still reachable after C release",
-    );
-
-    // Sanity: B's daemon is still up; suppress an unused-binding lint.
-    void configB;
-  },
-);
-
-// Concurrent agent-ring collection: A invites B, B invites C, C
-// invites A. After every party releases their pet name and the
-// matching '@pins/guest-*' entry, every daemon's local guest-handle
-// formula (the pin) should collect.
-//
-// This is the canonical "no central authority" GC case: each daemon
-// independently holds its local handle for one neighbour by way of a
-// pin; the ring is only realized through pet-name edges. No remote
-// retention edge should keep any local handle alive once both names
-// on its own daemon are gone.
-test.serial(
-  'agent ring (A->B->C->A) collects after all roots released',
-  async t => {
-    t.timeout(60000);
-    const { host: hostA, config: configA } =
-      await prepareHostWithGcAndNetwork(t);
-    const { host: hostB, config: configB } =
-      await prepareHostWithGcAndNetwork(t);
-    const { host: hostC, config: configC } =
-      await prepareHostWithGcAndNetwork(t);
-
-    // A -> B
-    const invAB = await E(hostA).invite('bob');
-    await E(hostB).accept(await E(invAB).locate(), 'alice');
-    // B -> C
-    const invBC = await E(hostB).invite('carol');
-    await E(hostC).accept(await E(invBC).locate(), 'bob');
-    // C -> A
-    const invCA = await E(hostC).invite('alice');
-    await E(hostA).accept(await E(invCA).locate(), 'carol');
-
-    // Capture each daemon's local guest-handle pin ids.  The pet
-    // names ('bob', 'carol', 'alice', ...) point to remote handles
-    // and are not in any local DB; only the @pins entries are.
-    const aBobPin = await E(hostA).identify('@pins', 'guest-bob');
-    const aCarolPin = await E(hostA).identify('@pins', 'guest-carol');
-    const bAlicePin = await E(hostB).identify('@pins', 'guest-alice');
-    const bCarolPin = await E(hostB).identify('@pins', 'guest-carol');
-    const cBobPin = await E(hostC).identify('@pins', 'guest-bob');
-    const cAlicePin = await E(hostC).identify('@pins', 'guest-alice');
-
-    // Sanity: every pin formula is currently reachable in its own DB.
-    for (const [label, statePath, id] of [
-      ['A bob pin', configA.statePath, aBobPin],
-      ['A carol pin', configA.statePath, aCarolPin],
-      ['B alice pin', configB.statePath, bAlicePin],
-      ['B carol pin', configB.statePath, bCarolPin],
-      ['C bob pin', configC.statePath, cBobPin],
-      ['C alice pin', configC.statePath, cAlicePin],
-    ]) {
-      t.true(formulaExistsInDb(statePath, id), `${label} exists pre-release`);
-    }
-
-    // Every party drops every reference (pet name + pin) for the ring.
-    await Promise.all([
-      E(hostA).remove('bob'),
-      E(hostA).remove('@pins', 'guest-bob'),
-      E(hostA).remove('carol'),
-      E(hostA).remove('@pins', 'guest-carol'),
-      E(hostB).remove('alice'),
-      E(hostB).remove('@pins', 'guest-alice'),
-      E(hostB).remove('carol'),
-      E(hostB).remove('@pins', 'guest-carol'),
-      E(hostC).remove('bob'),
-      E(hostC).remove('@pins', 'guest-bob'),
-      E(hostC).remove('alice'),
-      E(hostC).remove('@pins', 'guest-alice'),
-    ]);
-
-    // Every local guest-handle pin formula should collect on its own
-    // daemon. We assert per-daemon so a single regression points at
-    // exactly which side leaked.
-    const allCollected = () =>
-      !formulaExistsInDb(configA.statePath, aBobPin) &&
-      !formulaExistsInDb(configA.statePath, aCarolPin) &&
-      !formulaExistsInDb(configB.statePath, bAlicePin) &&
-      !formulaExistsInDb(configB.statePath, bCarolPin) &&
-      !formulaExistsInDb(configC.statePath, cBobPin) &&
-      !formulaExistsInDb(configC.statePath, cAlicePin);
-
-    await waitForCondition(allCollected, { timeoutMs: 15000 });
-
-    for (const [label, statePath, id] of [
-      ['A bob pin', configA.statePath, aBobPin],
-      ['A carol pin', configA.statePath, aCarolPin],
-      ['B alice pin', configB.statePath, bAlicePin],
-      ['B carol pin', configB.statePath, bCarolPin],
-      ['C bob pin', configC.statePath, cBobPin],
-      ['C alice pin', configC.statePath, cAlicePin],
-    ]) {
-      t.false(formulaExistsInDb(statePath, id), `${label} collected`);
-    }
-  },
-);
+runMultiplayerSuite({ test, network: tcpNetwork });
